# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Insight Exam API - Educational platform for analyzing past exam questions
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface User {
  id: number;
  name: string;
  email: string;
  createdAt: string;
}

export interface RegisterBody {
  /** @minLength 2 */
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
}

export interface LoginBody {
  email: string;
  password: string;
}

export interface AuthResponse {
  user: User;
  message: string;
}

export interface Course {
  id: number;
  userId: number;
  name: string;
  code: string;
  professor: string;
  questionCount: number;
  createdAt: string;
}

export interface CreateCourseBody {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  code: string;
  /** @minLength 1 */
  professor: string;
}

export type QuestionQuestionType =
  (typeof QuestionQuestionType)[keyof typeof QuestionQuestionType];

export const QuestionQuestionType = {
  multiple_choice: "multiple_choice",
  true_false: "true_false",
  essay: "essay",
  fill_blank: "fill_blank",
} as const;

export type QuestionDifficulty =
  (typeof QuestionDifficulty)[keyof typeof QuestionDifficulty];

export const QuestionDifficulty = {
  easy: "easy",
  medium: "medium",
  hard: "hard",
} as const;

export interface Question {
  id: number;
  courseId: number;
  text: string;
  chapter: string;
  questionType: QuestionQuestionType;
  difficulty: QuestionDifficulty;
  examPeriod: string;
  createdAt: string;
}

export type CreateQuestionBodyQuestionType =
  (typeof CreateQuestionBodyQuestionType)[keyof typeof CreateQuestionBodyQuestionType];

export const CreateQuestionBodyQuestionType = {
  multiple_choice: "multiple_choice",
  true_false: "true_false",
  essay: "essay",
  fill_blank: "fill_blank",
} as const;

export type CreateQuestionBodyDifficulty =
  (typeof CreateQuestionBodyDifficulty)[keyof typeof CreateQuestionBodyDifficulty];

export const CreateQuestionBodyDifficulty = {
  easy: "easy",
  medium: "medium",
  hard: "hard",
} as const;

export interface CreateQuestionBody {
  /** @minLength 1 */
  text: string;
  /** @minLength 1 */
  chapter: string;
  questionType: CreateQuestionBodyQuestionType;
  difficulty: CreateQuestionBodyDifficulty;
  /** @minLength 1 */
  examPeriod: string;
}

export interface ChapterFrequency {
  chapter: string;
  count: number;
  percentage: number;
}

export interface QuestionTypeDistribution {
  questionType: string;
  count: number;
  percentage: number;
}

export interface DifficultyDistribution {
  difficulty: string;
  count: number;
  percentage: number;
}

export interface YearlyFrequency {
  examPeriod: string;
  count: number;
}

export interface CourseAnalytics {
  totalQuestions: number;
  mostRepeatedChapter: string;
  mostCommonType: string;
  dominantDifficulty: string;
  chapterFrequency: ChapterFrequency[];
  typeDistribution: QuestionTypeDistribution[];
  difficultyDistribution: DifficultyDistribution[];
  yearlyFrequency: YearlyFrequency[];
  recommendations: string[];
}

export interface DashboardSummary {
  totalCourses: number;
  totalQuestions: number;
  recentCourses: Course[];
}