# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Insight Exam API - Educational platform for analyzing past exam questions
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication operations
  - name: courses
    description: Course management
  - name: questions
    description: Question management
  - name: analytics
    description: Analytics and dashboard data
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: User registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout user
      responses:
        "200":
          description: Logout successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /courses:
    get:
      operationId: getCourses
      tags: [courses]
      summary: Get all courses for the current user
      responses:
        "200":
          description: List of courses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Course"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createCourse
      tags: [courses]
      summary: Create a new course
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCourseBody"
      responses:
        "201":
          description: Course created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /courses/{id}:
    get:
      operationId: getCourse
      tags: [courses]
      summary: Get a single course by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Course details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
        "404":
          description: Course not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteCourse
      tags: [courses]
      summary: Delete a course
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  /courses/{id}/questions:
    get:
      operationId: getCourseQuestions
      tags: [questions]
      summary: Get all questions for a course
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of questions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Question"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: addQuestion
      tags: [questions]
      summary: Add a question to a course
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateQuestionBody"
      responses:
        "201":
          description: Question created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /courses/{id}/analytics:
    get:
      operationId: getCourseAnalytics
      tags: [analytics]
      summary: Get analytics for a course
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Analytics data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourseAnalytics"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Course not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [analytics]
      summary: Get overall dashboard summary for the current user
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message

    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - createdAt

    RegisterBody:
      type: object
      properties:
        name:
          type: string
          minLength: 2
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
      required:
        - name
        - email
        - password

    LoginBody:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - email
        - password

    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        message:
          type: string
      required:
        - user
        - message

    Course:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        code:
          type: string
        professor:
          type: string
        questionCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - name
        - code
        - professor
        - questionCount
        - createdAt

    CreateCourseBody:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        code:
          type: string
          minLength: 1
        professor:
          type: string
          minLength: 1
      required:
        - name
        - code
        - professor

    Question:
      type: object
      properties:
        id:
          type: integer
        courseId:
          type: integer
        text:
          type: string
        chapter:
          type: string
        questionType:
          type: string
          enum:
            - multiple_choice
            - true_false
            - essay
            - fill_blank
        difficulty:
          type: string
          enum:
            - easy
            - medium
            - hard
        examPeriod:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - courseId
        - text
        - chapter
        - questionType
        - difficulty
        - examPeriod
        - createdAt

    CreateQuestionBody:
      type: object
      properties:
        text:
          type: string
          minLength: 1
        chapter:
          type: string
          minLength: 1
        questionType:
          type: string
          enum:
            - multiple_choice
            - true_false
            - essay
            - fill_blank
        difficulty:
          type: string
          enum:
            - easy
            - medium
            - hard
        examPeriod:
          type: string
          minLength: 1
      required:
        - text
        - chapter
        - questionType
        - difficulty
        - examPeriod

    ChapterFrequency:
      type: object
      properties:
        chapter:
          type: string
        count:
          type: integer
        percentage:
          type: number
      required:
        - chapter
        - count
        - percentage

    QuestionTypeDistribution:
      type: object
      properties:
        questionType:
          type: string
        count:
          type: integer
        percentage:
          type: number
      required:
        - questionType
        - count
        - percentage

    DifficultyDistribution:
      type: object
      properties:
        difficulty:
          type: string
        count:
          type: integer
        percentage:
          type: number
      required:
        - difficulty
        - count
        - percentage

    YearlyFrequency:
      type: object
      properties:
        examPeriod:
          type: string
        count:
          type: integer
      required:
        - examPeriod
        - count

    CourseAnalytics:
      type: object
      properties:
        totalQuestions:
          type: integer
        mostRepeatedChapter:
          type: string
        mostCommonType:
          type: string
        dominantDifficulty:
          type: string
        chapterFrequency:
          type: array
          items:
            $ref: "#/components/schemas/ChapterFrequency"
        typeDistribution:
          type: array
          items:
            $ref: "#/components/schemas/QuestionTypeDistribution"
        difficultyDistribution:
          type: array
          items:
            $ref: "#/components/schemas/DifficultyDistribution"
        yearlyFrequency:
          type: array
          items:
            $ref: "#/components/schemas/YearlyFrequency"
        recommendations:
          type: array
          items:
            type: string
      required:
        - totalQuestions
        - mostRepeatedChapter
        - mostCommonType
        - dominantDifficulty
        - chapterFrequency
        - typeDistribution
        - difficultyDistribution
        - yearlyFrequency
        - recommendations

    DashboardSummary:
      type: object
      properties:
        totalCourses:
          type: integer
        totalQuestions:
          type: integer
        recentCourses:
          type: array
          items:
            $ref: "#/components/schemas/Course"
      required:
        - totalCourses
        - totalQuestions
        - recentCourses