# lib/api-zod/src/generated/types/chapterFrequency.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Insight Exam API - Educational platform for analyzing past exam questions
 * OpenAPI spec version: 0.1.0
 */

export interface ChapterFrequency {
  chapter: string;
  count: number;
  percentage: number;
}