# lib/api-zod/src/generated/types/createQuestionBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Insight Exam API - Educational platform for analyzing past exam questions
 * OpenAPI spec version: 0.1.0
 */
import type { CreateQuestionBodyDifficulty } from "./createQuestionBodyDifficulty";
import type { CreateQuestionBodyQuestionType } from "./createQuestionBodyQuestionType";

export interface CreateQuestionBody {
  /** @minLength 1 */
  text: string;
  /** @minLength 1 */
  chapter: string;
  questionType: CreateQuestionBodyQuestionType;
  difficulty: CreateQuestionBodyDifficulty;
  /** @minLength 1 */
  examPeriod: string;
}