# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Insight Exam API - Educational platform for analyzing past exam questions
 * OpenAPI spec version: 0.1.0
 */
import type { Course } from "./course";

export interface DashboardSummary {
  totalCourses: number;
  totalQuestions: number;
  recentCourses: Course[];
}