# lib/api-zod/src/generated/types/question.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Insight Exam API - Educational platform for analyzing past exam questions
 * OpenAPI spec version: 0.1.0
 */
import type { QuestionDifficulty } from "./questionDifficulty";
import type { QuestionQuestionType } from "./questionQuestionType";

export interface Question {
  id: number;
  courseId: number;
  text: string;
  chapter: string;
  questionType: QuestionQuestionType;
  difficulty: QuestionDifficulty;
  examPeriod: string;
  createdAt: Date;
}